+++ MOP/src/Misc/CustomExtensions.cs
@@ -61,5 +61,24 @@ namespace MOP
             // Nothing has been found? Return false.
             return false;
         }
+
+        static string[] rainbow = new string[] { "red", "orange", "yellow", "green", "blue", "purple" };
+        public static string Rainbowmize(this string input)
+        {
+            char[] inputArray = input.ToCharArray();
+            string output = "";
+            int colorNumber = 0;
+            for (int i = 0; i < inputArray.Length; i++)
+            {
+                if (colorNumber >= rainbow.Length)
+                    colorNumber = 0;
+
+                string color = rainbow[colorNumber];
+                output += $"<color={color}>{inputArray[i]}</color>";
+                colorNumber++;
+            }
+
+            return output;
+        }
     }
 }